+++ CHANGELOG.md
@@ -2,6 +2,30 @@
 
 This changelog covers all three packges, as they are (for now) updated as a whole
 
+## v0.0.12
+
+### atomic-data-browser
+
+- Fix tests
+- Cleaned up Resource form #51
+- Handle usages left in Invites #45
+- Add social meta tags #44
+- Add fetch as JSON / JSON-AD / Turtle and more to data pages
+- Fix bug with invites
+- Various styling improvements
+- Add Atomic Data Logo
+- Dark mode syncs with user
+- Scroll to top on page change #47
+- Improve keyboard shortcuts for edit / data view #52
+
+### @tomic/react
+
+- Resources will update when properties change
+
+### @tomic/lib
+
+- Add `getCommitBuilder` and `hasChanges` function to `resource` and `commitBuilder`
+
 ## v0.0.11
 
 - Split packages, switch to monorepo